Completed Payment Workflow, + Wishlist & Comments. Signed-off-by: Ruth Nunez <ruth.nunez@teamcodeflux.com>

+++ grails-app/domain/com/productHub/domain/Comment.groovy
@@ -6,7 +6,7 @@ class Comment {
 	Product product
 	Store store
 	String remarks
-	float rating
+	int rating
 
     static constraints = {
 		product(nullable:true)
